+++ src/unity/Runtime/AppsFlyer/IAppsFlyer.cs
@@ -1,0 +1,9 @@
+namespace EE {
+    public interface IAppsFlyer : IPlugin {
+        void Initialize(string devKey, string iosAppId);
+        void StartTracking();
+        string GetDeviceId();
+        void SetDebugEnabled(bool enabled);
+        void SetStopTracking(bool enabled);
+    }
+}